moved more code from RedisQues
